Move PrivacyFeature enum into separate file
